refactor(benchmark): consolidate into single script with markdown reports
- Move ComparativeEvaluator from agents/benchmark.py to scripts/benchmark_agents.py
- Add markdown report generation (evaluation-report-{date}-{time}-{commit}.md)
- Delete redundant agents/benchmark.py and tests/e2e/test_benchmark.py
- Add reports/ directory with .gitignore for generated files
- Add --no-report CLI flag for skipping report generation

Benchmark results (error-analysis scenario):
- CrewAI: 49s, 4/5 (Winner)
- AutoGen: 44s, 3/5

Note: OpenHands agent times out due to tool-based exploration; use autogen/crewai for benchmarks
