+++ scripts/benchmark_agents.py
@@ -3,6 +3,7 @@
 Consolidated Agent Benchmark Script
 
 Benchmarks VibeTeam agents across multiple scenarios and frameworks.
+Generates markdown evaluation reports.
 
 Usage:
     python scripts/benchmark_agents.py --list                    # List available scenarios
@@ -15,21 +16,57 @@ Scenarios:
     sentry-summary  - Summarize Sentry issues for the week
     github-triage   - Triage GitHub issues with labels and priority
     release-notes   - Generate release notes from merged PRs
+
+Reports:
+    Generated in reports/ directory with format: evaluation-report-{date}-{time}-{commit}.md
 """
 
 import argparse
 import asyncio
 import json
 import os
+import re
+import subprocess
 import sys
 import time
-from dataclasses import dataclass
+from dataclasses import asdict, dataclass
+from datetime import datetime, timezone
+from pathlib import Path
 from typing import Any
+
+import httpx
 
 # Add parent directory to path for imports
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from agents.benchmark import ComparativeEvaluator, ComparativeResult
+
+# ==============================================================================
+# Configuration
+# ==============================================================================
+
+
+class BenchmarkConfig:
+    """Benchmark configuration from environment."""
+
+    AZURE_API_KEY = os.getenv("AZURE_OPENAI_API_KEY", os.getenv("AZURE_API_KEY", ""))
+    AZURE_API_BASE = os.getenv("AZURE_OPENAI_ENDPOINT", os.getenv("AZURE_API_BASE", ""))
+    AZURE_API_VERSION = os.getenv("AZURE_API_VERSION", "2024-08-01-preview")
+    JUDGE_MODEL = os.getenv(
+        "BENCHMARK_JUDGE_MODEL", os.getenv("AZURE_OPENAI_DEPLOYMENT", "gpt-5-2")
+    )
+    REPORTS_DIR = Path(os.getenv("BENCHMARK_REPORTS_DIR", "reports"))
+
+    @classmethod
+    def get_azure_api_base(cls) -> str:
+        """Get AZURE_API_BASE with protocol validation."""
+        base = cls.AZURE_API_BASE
+        if not base:
+            raise ValueError("AZURE_API_BASE environment variable not set")
+        if not base.startswith(("http://", "https://")):
+            raise ValueError(
+                f"AZURE_API_BASE must start with 'http://' or 'https://'. Got: '{base}'"
+            )
+        return base.rstrip("/")
 
 
 # ==============================================================================
@@ -168,6 +205,221 @@ class FrameworkResult:
     latency_ms: int
     success: bool
     error: str | None = None
+
+
+@dataclass
+class ComparativeScore:
+    """Score for a single agent from comparative evaluation."""
+
+    framework: str
+    score: int  # 0-5 scale
+    feedback: str
+
+    def to_dict(self) -> dict[str, Any]:
+        return asdict(self)
+
+
+@dataclass
+class ComparativeResult:
+    """Result from comparative LLM-as-judge evaluation."""
+
+    task: str
+    scores: dict[str, ComparativeScore]  # framework -> score
+    winner: str
+    reasoning: str
+    judge_model: str
+    evaluation_time_ms: int = 0
+
+    def to_dict(self) -> dict[str, Any]:
+        return {
+            "task": self.task[:200],
+            "scores": {k: v.to_dict() for k, v in self.scores.items()},
+            "winner": self.winner,
+            "reasoning": self.reasoning,
+            "judge_model": self.judge_model,
+            "evaluation_time_ms": self.evaluation_time_ms,
+        }
+
+    def __str__(self) -> str:
+        lines = [
+            "=" * 60,
+            "LLM-AS-JUDGE EVALUATION RESULTS",
+            "=" * 60,
+            "",
+        ]
+        for fw, score in sorted(self.scores.items(), key=lambda x: x[1].score, reverse=True):
+            lines.append(f"{fw.upper()}: {score.score}/5")
+            lines.append(f"  Feedback: {score.feedback}")
+            lines.append("")
+
+        lines.extend(
+            [
+                "-" * 60,
+                f"WINNER: {self.winner.upper()}",
+                f"Reasoning: {self.reasoning}",
+                "-" * 60,
+                f"Judge: {self.judge_model} | Time: {self.evaluation_time_ms}ms",
+            ]
+        )
+        return "\n".join(lines)
+
+
+# ==============================================================================
+# Comparative Evaluator (LLM-as-Judge)
+# ==============================================================================
+
+
+class ComparativeEvaluator:
+    """
+    Evaluates multiple agent responses side-by-side using LLM-as-judge.
+
+    Uses a simple 0-5 scoring scale:
+    - 0: Failed completely or error
+    - 1: Attempted but mostly wrong
+    - 2: Partially correct, missing key elements
+    - 3: Acceptable, addresses main points
+    - 4: Good, comprehensive and accurate
+    - 5: Excellent, exceeds expectations
+    """
+
+    COMPARATIVE_PROMPT = """You are an expert evaluator comparing AI agent responses to the same task.
+
+TASK:
+{task}
+
+AGENT RESPONSES:
+
+=== AUTOGEN ===
+{autogen_response}
+
+=== CREWAI ===
+{crewai_response}
+
+=== OPENHANDS ===
+{openhands_response}
+
+Score each agent from 0-5 based on how well they completed the task:
+- 0: Failed completely, error, or refused to answer
+- 1: Attempted but mostly wrong or unhelpful
+- 2: Partially correct but missing key elements
+- 3: Acceptable, addresses the main points adequately
+- 4: Good, comprehensive and accurate response
+- 5: Excellent, exceeds expectations with actionable insights
+
+Consider:
+- Accuracy: Is the information correct and not hallucinated?
+- Completeness: Does it address all parts of the task?
+- Usefulness: Is the response actionable and helpful?
+- Clarity: Is it well-organized and easy to understand?
+
+Return ONLY valid JSON in this exact format:
+{{
+  "autogen": {{"score": 0, "feedback": "Brief explanation"}},
+  "crewai": {{"score": 0, "feedback": "Brief explanation"}},
+  "openhands": {{"score": 0, "feedback": "Brief explanation"}},
+  "winner": "framework_name",
+  "reasoning": "One sentence explaining why this framework won"
+}}"""
+
+    def __init__(self, config: BenchmarkConfig | None = None):
+        self.config = config or BenchmarkConfig()
+
+    async def evaluate(
+        self,
+        task: str,
+        responses: dict[str, str],
+    ) -> ComparativeResult:
+        """
+        Evaluate multiple agent responses side-by-side.
+
+        Args:
+            task: The original task/prompt given to agents
+            responses: Dict mapping framework name to response text
+
+        Returns:
+            ComparativeResult with scores for each framework
+        """
+        start_time = time.perf_counter()
+
+        # Ensure we have all three frameworks (use empty string if missing)
+        autogen_resp = responses.get("autogen", "(No response)")[:3000]
+        crewai_resp = responses.get("crewai", "(No response)")[:3000]
+        openhands_resp = responses.get("openhands", "(No response)")[:3000]
+
+        prompt = self.COMPARATIVE_PROMPT.format(
+            task=task[:1000],
+            autogen_response=autogen_resp,
+            crewai_response=crewai_resp,
+            openhands_response=openhands_resp,
+        )
+
+        try:
+            result_json = await self._call_llm(prompt)
+            result = self._parse_result(result_json)
+            result.task = task[:200]
+            result.judge_model = self.config.JUDGE_MODEL
+            result.evaluation_time_ms = int((time.perf_counter() - start_time) * 1000)
+            return result
+        except Exception as e:
+            # Return error result
+            return ComparativeResult(
+                task=task[:200],
+                scores={
+                    fw: ComparativeScore(framework=fw, score=0, feedback=f"Evaluation error: {e}")
+                    for fw in ["autogen", "crewai", "openhands"]
+                },
+                winner="none",
+                reasoning=f"Evaluation failed: {e}",
+                judge_model=self.config.JUDGE_MODEL,
+                evaluation_time_ms=int((time.perf_counter() - start_time) * 1000),
+            )
+
+    async def _call_llm(self, prompt: str) -> str:
+        """Call Azure OpenAI for comparative evaluation."""
+        api_base = self.config.get_azure_api_base()
+        async with httpx.AsyncClient(timeout=90.0) as client:
+            response = await client.post(
+                f"{api_base}/openai/deployments/{self.config.JUDGE_MODEL}/chat/completions",
+                headers={
+                    "api-key": self.config.AZURE_API_KEY,
+                    "Content-Type": "application/json",
+                },
+                params={"api-version": self.config.AZURE_API_VERSION},
+                json={
+                    "messages": [{"role": "user", "content": prompt}],
+                    "temperature": 0.1,
+                    "max_completion_tokens": 800,
+                },
+            )
+            response.raise_for_status()
+            data = response.json()
+            return data["choices"][0]["message"]["content"]
+
+    def _parse_result(self, json_str: str) -> ComparativeResult:
+        """Parse JSON result from LLM response."""
+        # Extract JSON from potential markdown code blocks
+        json_match = re.search(r"\{[\s\S]*\}", json_str)
+        if not json_match:
+            raise ValueError("No JSON found in LLM response")
+
+        data = json.loads(json_match.group())
+
+        scores = {}
+        for framework in ["autogen", "crewai", "openhands"]:
+            fw_data = data.get(framework, {})
+            scores[framework] = ComparativeScore(
+                framework=framework,
+                score=int(fw_data.get("score", 0)),
+                feedback=str(fw_data.get("feedback", "No feedback")),
+            )
+
+        return ComparativeResult(
+            task="",
+            scores=scores,
+            winner=str(data.get("winner", "unknown")),
+            reasoning=str(data.get("reasoning", "No reasoning provided")),
+            judge_model="",
+        )
 
 
 # ==============================================================================
@@ -310,6 +562,154 @@ def print_comparison_table(results: list[FrameworkResult]) -> None:
 
 
 # ==============================================================================
+# Markdown Report Generation
+# ==============================================================================
+
+
+def get_git_commit() -> str:
+    """Get current git commit hash."""
+    try:
+        result = subprocess.run(
+            ["git", "rev-parse", "--short", "HEAD"],
+            capture_output=True,
+            text=True,
+            timeout=5,
+        )
+        return result.stdout.strip() if result.returncode == 0 else "unknown"
+    except Exception:
+        return "unknown"
+
+
+def generate_markdown_report(
+    scenario_id: str,
+    scenario: dict[str, Any],
+    results: list[FrameworkResult],
+    eval_result: ComparativeResult,
+) -> str:
+    """Generate a markdown evaluation report."""
+    now = datetime.now(timezone.utc)
+    commit = get_git_commit()
+
+    lines = [
+        f"# Agent Benchmark Evaluation Report",
+        "",
+        f"**Generated**: {now.strftime('%Y-%m-%d %H:%M:%S UTC')}",
+        f"**Commit**: `{commit}`",
+        f"**Scenario**: {scenario['name']} (`{scenario_id}`)",
+        f"**Role**: {scenario['role']}",
+        "",
+        "---",
+        "",
+        "## Summary",
+        "",
+        "| Framework | Status | Latency | Score | Response Length |",
+        "|-----------|--------|---------|-------|-----------------|",
+    ]
+
+    for r in sorted(
+        results,
+        key=lambda x: -eval_result.scores.get(
+            x.framework, ComparativeScore(x.framework, 0, "")
+        ).score,
+    ):
+        score = eval_result.scores.get(r.framework, ComparativeScore(r.framework, 0, "N/A"))
+        status = "PASS" if r.success else "FAIL"
+        lines.append(
+            f"| {r.framework.upper()} | {status} | {r.latency_ms}ms | {score.score}/5 | {len(r.response)} chars |"
+        )
+
+    lines.extend(
+        [
+            "",
+            f"**Winner**: {eval_result.winner.upper()}",
+            "",
+            f"**Reasoning**: {eval_result.reasoning}",
+            "",
+            "---",
+            "",
+            "## Task",
+            "",
+            "```",
+            scenario["task"],
+            "```",
+            "",
+            "---",
+            "",
+        ]
+    )
+
+    # Add each framework's response
+    for r in results:
+        score = eval_result.scores.get(r.framework, ComparativeScore(r.framework, 0, "N/A"))
+        lines.extend(
+            [
+                f"## {r.framework.upper()} Response",
+                "",
+                f"**Score**: {score.score}/5",
+                f"**Feedback**: {score.feedback}",
+                f"**Latency**: {r.latency_ms}ms",
+                "",
+            ]
+        )
+
+        if r.error:
+            lines.extend(
+                [
+                    f"**Error**: {r.error}",
+                    "",
+                ]
+            )
+
+        if r.response:
+            lines.extend(
+                [
+                    "### Output",
+                    "",
+                    "```",
+                    r.response,
+                    "```",
+                    "",
+                ]
+            )
+        else:
+            lines.append("*(No response)*\n")
+
+        lines.append("---\n")
+
+    # Evaluation metadata
+    lines.extend(
+        [
+            "## Evaluation Metadata",
+            "",
+            f"- **Judge Model**: {eval_result.judge_model}",
+            f"- **Evaluation Time**: {eval_result.evaluation_time_ms}ms",
+            "",
+        ]
+    )
+
+    return "\n".join(lines)
+
+
+def save_markdown_report(content: str, scenario_id: str) -> Path:
+    """Save markdown report to file."""
+    now = datetime.now(timezone.utc)
+    commit = get_git_commit()
+
+    # Create reports directory
+    reports_dir = BenchmarkConfig.REPORTS_DIR
+    reports_dir.mkdir(parents=True, exist_ok=True)
+
+    # Generate filename: evaluation-report-{date}-{time}-{commit}.md
+    filename = f"evaluation-report-{now.strftime('%Y%m%d')}-{now.strftime('%H%M%S')}-{commit}.md"
+    filepath = reports_dir / filename
+
+    with open(filepath, "w") as f:
+        f.write(content)
+
+    return filepath
+
+
+# ==============================================================================
 # Benchmark Runner
 # ==============================================================================
 
@@ -319,6 +719,7 @@ async def run_scenario(
     frameworks: list[str],
     timeout: float,
     output_json: bool = False,
+    save_report: bool = True,
 ) -> dict[str, Any]:
     """Run a single scenario across all frameworks."""
     scenario = SCENARIOS[scenario_id]
@@ -357,6 +758,12 @@ async def run_scenario(
     if not output_json:
         print(str(eval_result))
 
+    # Generate and save markdown report
+    if save_report:
+        report_content = generate_markdown_report(scenario_id, scenario, results, eval_result)
+        report_path = save_markdown_report(report_content, scenario_id)
+        print(f"\n>>> Report saved: {report_path}")
+
     # Prepare output
     output = {
         "scenario": scenario_id,
@@ -369,6 +776,7 @@ async def run_scenario(
                 "success": r.success,
                 "latency_ms": r.latency_ms,
                 "response_length": len(r.response),
+                "response": r.response,
                 "error": r.error,
             }
             for r in results
@@ -390,13 +798,14 @@ async def run_all_scenarios(
     frameworks: list[str],
     timeout: float,
     output_json: bool = False,
+    save_report: bool = True,
 ) -> dict[str, Any]:
     """Run all scenarios and produce aggregate results."""
     all_results = {}
     winners = {}
 
     for scenario_id in SCENARIOS:
-        result = await run_scenario(scenario_id, frameworks, timeout, output_json)
+        result = await run_scenario(scenario_id, frameworks, timeout, output_json, save_report)
         all_results[scenario_id] = result
         winners[scenario_id] = result["winner"]
 
@@ -409,7 +818,7 @@ async def run_all_scenarios(
             print(f"{scenario_id:<20} {winner.upper():<15}")
 
         # Count wins
-        win_counts = {}
+        win_counts: dict[str, int] = {}
         for winner in winners.values():
             win_counts[winner] = win_counts.get(winner, 0) + 1
 
@@ -457,6 +866,7 @@ Examples:
   python scripts/benchmark_agents.py --scenario sentry-summary --timeout 300
   python scripts/benchmark_agents.py --all --json
   python scripts/benchmark_agents.py --scenario github-triage --frameworks autogen crewai
+  python scripts/benchmark_agents.py --scenario error-analysis --no-report
         """,
     )
     parser.add_argument(
@@ -493,6 +903,11 @@ Examples:
         action="store_true",
         help="Output results as JSON only",
     )
+    parser.add_argument(
+        "--no-report",
+        action="store_true",
+        help="Skip generating markdown report",
+    )
 
     args = parser.parse_args()
 
@@ -501,11 +916,15 @@ Examples:
         return
 
     if args.all:
-        await run_all_scenarios(args.frameworks, args.timeout, args.json)
+        await run_all_scenarios(
+            args.frameworks, args.timeout, args.json, save_report=not args.no_report
+        )
         return
 
     if args.scenario:
-        await run_scenario(args.scenario, args.frameworks, args.timeout, args.json)
+        await run_scenario(
+            args.scenario, args.frameworks, args.timeout, args.json, save_report=not args.no_report
+        )
         return
 
     # No action specified, show help
